[refactor] apply cargo clippy

+++ src/cpu/rv32/execution/inst_16/c_extension.rs
@@ -113,7 +113,7 @@ pub fn exec(inst: &Instruction, cpu: &mut Cpu32) -> Result<(), (Option<u32>, Tra
             let next_pc = cpu.pc + INST_SIZE;
             // setting the least-significant bit of
             // the result to zero                ->vvvvvv
-            cpu.update_pc((cpu.regs.read(inst.rs1) & !0x1) as u32);
+            cpu.update_pc(cpu.regs.read(inst.rs1) & !0x1);
             cpu.regs.write(Some(REG_LINK), next_pc);
         }
         OpecodeKind::OP_C_BEQZ => {
@@ -127,7 +127,7 @@ pub fn exec(inst: &Instruction, cpu: &mut Cpu32) -> Result<(), (Option<u32>, Tra
             }
         }
         OpecodeKind::OP_C_JR => {
-            cpu.update_pc(cpu.regs.read(inst.rs1) as u32);
+            cpu.update_pc(cpu.regs.read(inst.rs1));
         }
         OpecodeKind::OP_C_MV => {
             cpu.regs.write(inst.rd, cpu.regs.read(inst.rs2));
