+++ src/cpu/rv32/execution/inst_32/m_extension.rs
@@ -19,13 +19,13 @@ pub fn exec(inst: &Instruction, cpu: &mut Cpu32) -> Result<(), (Option<u32>, Tra
         OpecodeKind::OP_MULHSU => {
             cpu.regs.write(
                 inst.rd,
-                ((rs1 as i32 as i64 * rs2 as u32 as u64 as i64) >> 32) as u32,
+                ((rs1 as i32 as i64 * rs2 as u64 as i64) >> 32) as u32,
             );
         }
         OpecodeKind::OP_MULHU => {
             cpu.regs.write(
                 inst.rd,
-                ((rs1 as u32 as u64 * rs2 as u32 as u64) >> 32) as u32,
+                ((rs1 as u64 * rs2 as u64) >> 32) as u32,
             );
         }
         OpecodeKind::OP_DIV => {
@@ -41,7 +41,7 @@ pub fn exec(inst: &Instruction, cpu: &mut Cpu32) -> Result<(), (Option<u32>, Tra
                 cpu.regs.write(inst.rd, (2i32.pow(32) - 1) as u32);
             } else {
                 cpu.regs
-                    .write(inst.rd, (rs1 as u32 as u64 / rs2 as u32 as u64) as u32);
+                    .write(inst.rd, (rs1 as u64 / rs2 as u64) as u32);
             }
         }
         OpecodeKind::OP_REM => {
