+++ src/cpu/rv32/execution/inst_32/zicsr_extension.rs
@@ -74,17 +74,17 @@ pub fn exec(inst: &Instruction, cpu: &mut Cpu32) -> Result<(), (Option<u32>, Tra
 
     match inst.opc {
         OpecodeKind::OP_CSRRW => {
-            let rs1 = cpu.regs.read(inst.rs1) as u32;
+            let rs1 = cpu.regs.read(inst.rs1);
             cpu.regs.write(inst.rd, cpu.csrs.read(inst.rs2)?);
             cpu.csrs.write(inst.rs2, rs1);
         }
         OpecodeKind::OP_CSRRS => {
-            let rs1 = cpu.regs.read(inst.rs1) as u32;
+            let rs1 = cpu.regs.read(inst.rs1);
             cpu.regs.write(inst.rd, cpu.csrs.read(inst.rs2)?);
             cpu.csrs.bitset(inst.rs2, rs1);
         }
         OpecodeKind::OP_CSRRC => {
-            let rs1 = cpu.regs.read(inst.rs1) as u32;
+            let rs1 = cpu.regs.read(inst.rs1);
             cpu.regs.write(inst.rd, cpu.csrs.read(inst.rs2)?);
             cpu.csrs.bitclr(inst.rs2, rs1);
         }
